Isolating startSpendFromLiquidMultisig function
